Ensure syntax error locations appear in backtraces
Errors which are thrown directly inside jl_toplevel_eval_flex do not get
an entry in the backtrace. The most prominent of these are syntax
errors, but there's other cases too.

Fix this by constructing a julia expression throwing the appropriate
error, and evaling it to generate a julia-level frame.